+++ src/Kernel.php
@@ -139,9 +139,10 @@ class Kernel extends BaseKernel implements CompilerPassInterface
 
         $container->setParameter('kernel.secret', Auth::privateKey());
 
-        $container->addResource(new FileResource(Setup::getConfigPath()));
         $container->addResource(new FileResource(Setup::getPrivateKeyPath()));
-        $container->addResource(new FileResource(Setup::getSetupFile()));
+        if (file_exists(Setup::getSetupFile())) {
+            $container->addResource(new FileResource(Setup::getSetupFile()));
+        }
         $container->addResource(new FileResource("{$resourceDir}/bundles.php"));
 
         $container->setParameter('container.dumper.inline_class_loader', true);