+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See [Upgrading] for details on how to upgrade.
 - Add `md_textarea.tpl.html` include to handle markdown enabled text areas, #989
 - Move "bordered" and "grid" css to separate file, #990
 - With this release, we reached [1000th] Pull Request! 💯
+- Fix setup for initial install, #1007
 
 [3.9.11]: https://github.com/eventum/eventum/compare/v3.9.10...master
 [1000th]: https://github.com/eventum/eventum/pull/1000